Resolve multi-target platform issue with imports
