+++ @planetarium/account-web3-secret-storage/src/Web3KeyStore.ts
@@ -17,145 +17,13 @@ import {
   type ImportableKeyStore,
   RawPrivateKey,
 } from "@planetarium/account";
-
-const IS_NODE = typeof window === "undefined";
-
-interface IFileSystem {
-  mkdir(path: string, options?: { recursive: true }): Promise<void>;
-  listFiles(directory: string): AsyncIterable<string>;
-  readFile(path: string, options?: { encoding: "utf8" }): Promise<string>;
-  delete(path: string): Promise<void>;
-  writeFile(path: string, content: string, encoding?: "utf8"): Promise<void>;
-}
-
-export class NodeJsFileSystem implements IFileSystem {
-  readonly #fs: typeof import("node:fs/promises");
-
-  private constructor(fs: typeof import("node:fs/promises")) {
-    this.#fs = fs;
-  }
-
-  static async create(): Promise<NodeJsFileSystem> {
-    return new NodeJsFileSystem(await import("node:fs/promises"));
-  }
-
-  async mkdir(path: string, options?: { recursive: true }): Promise<void> {
-    await this.#fs.mkdir(path, options);
-  }
-
-  readFile(path: string, options?: { encoding: "utf8" }): Promise<string> {
-    return this.#fs.readFile(path, options) as Promise<string>;
-  }
-
-  delete(path: string): Promise<void> {
-    return this.#fs.unlink(path);
-  }
-
-  async *listFiles(directory: string): AsyncIterable<string> {
-    let dir;
-    try {
-      dir = await this.#fs.opendir(directory);
-    } catch (e) {
-      if (
-        typeof e === "object" &&
-        e != null &&
-        "code" in e &&
-        e.code === "ENOENT"
-      ) {
-        // In case where there is no directory at all (it's likely the first
-        // time to run this operation in a system), it should be considered
-        // it's just empty (instead of considering it an exceptional case).
-        return;
-      }
-      throw e;
-    }
-    for await (const dirEntry of dir) {
-      if (!dirEntry.isFile()) continue;
-      yield dirEntry.name;
-    }
-  }
-
-  writeFile(path: string, content: string, encoding?: "utf8"): Promise<void> {
-    return this.#fs.writeFile(path, content, encoding);
-  }
-}
-
-export class LocalStorageFileSystem implements IFileSystem {
-  async mkdir(path: string, options?: unknown): Promise<void> {}
-
-  readFile(path: string, options?: { encoding: "utf8" }): Promise<string> {
-    const item = localStorage.getItem(path);
-    if (item == null) {
-      throw new Error("Not found");
-    }
-
-    return Promise.resolve(item);
-  }
-
-  async delete(path: string): Promise<void> {
-    localStorage.removeItem(path);
-  }
-
-  async *listFiles(directory: string): AsyncIterable<string> {
-    for (let i = 0; i < localStorage.length; ++i) {
-      const item = localStorage.key(i);
-      if (item == null) {
-        throw new Error(`Expected ${i}th item in localStorage.`);
-      }
-
-      if (item.startsWith(directory)) {
-        yield item.slice(directory.length);
-      }
-    }
-  }
-
-  async writeFile(
-    path: string,
-    content: string,
-    encoding?: "utf8",
-  ): Promise<void> {
-    localStorage.setItem(path, content);
-  }
-}
+import { join, getDefaultWeb3KeyStorePath } from "#path";
+export { getDefaultWeb3KeyStorePath } from "#path";
+import * as fs from "#fs";
 
 export interface Web3KeyStoreOptions {
   path?: string;
   passphraseEntry: PassphraseEntry;
-  fileSystem: IFileSystem;
-}
-
-function pathJoin(x: string, y: string) {
-  if (IS_NODE) {
-    return require("path").join(x, y);
-  }
-
-  return `${x}/${y}`;
-}
-
-/**
- * Determines the default key store path.  It depends on the platform:
- *
- * - Linux/macOS: `$HOME/.config/planetarium/keystore`
- * - Windows: `%AppData%\planetarium\keystore`
- */
-export function getDefaultWeb3KeyStorePath(): string {
-  if (IS_NODE) {
-    const { homedir } = require("node:os");
-    const path = require("node:path");
-    const baseDir =
-      process.platform === "win32"
-        ? process.env.AppData || path.join(homedir(), "AppData", "Roaming")
-        : process.env.XDG_CONFIG_HOME || path.join(homedir(), ".config");
-    // Note that it's not necessary to explicitly choose one of `path.win32` or
-    // `path.posix` here, but it makes unit tests less dependent on mocks:
-    return (process.platform === "win32" ? path.win32 : path.posix).join(
-      baseDir,
-      "planetarium",
-      "keystore",
-    );
-  } else {
-    return "/planetarium/account/web3-secret-storage/keystore";
-  }
 }
 
 const pattern =
@@ -183,19 +51,17 @@ export class Web3KeyStore
 {
   readonly #passphraseEntry: PassphraseEntry;
   readonly #accountOptions: Partial<Web3AccountOptions>;
-  readonly #fileSystem: IFileSystem;
 
   readonly path: string;
 
   constructor(options: Web3KeyStoreOptions & Partial<Web3AccountOptions>) {
     this.path = options.path || getDefaultWeb3KeyStorePath();
     this.#passphraseEntry = options.passphraseEntry;
-    this.#fileSystem = options.fileSystem;
     this.#accountOptions = options;
   }
 
   async *#listKeyFiles(): AsyncIterable<string> {
-    for await (const name of this.#fileSystem.listFiles(this.path)) {
+    for await (const name of fs.listFiles(this.path)) {
       yield name;
     }
   }
@@ -206,7 +72,7 @@ export class Web3KeyStore
     for await (const name of this.#listKeyFiles()) {
       const parsed = parseKeyFilename(name);
       if (parsed != null && parsed.keyId === keyId) {
-        return { ...parsed, path: pathJoin(this.path, name) };
+        return { ...parsed, path: join(this.path, name) };
       }
     }
     return undefined;
@@ -216,11 +82,11 @@ export class Web3KeyStore
     for await (const name of this.#listKeyFiles()) {
       const parsed = parseKeyFilename(name);
       if (parsed == null) continue;
-      const keyPath = pathJoin(this.path, name);
+      const keyPath = join(this.path, name);
       let json: unknown;
       try {
         json = JSON.parse(
-          await this.#fileSystem.readFile(keyPath, { encoding: "utf8" }),
+          await fs.readFile(keyPath, { encoding: "utf8" }),
         );
       } catch (_) {
         continue;
@@ -244,7 +110,7 @@ export class Web3KeyStore
     if (keyPath == null) return { result: "keyNotFound", keyId };
     let json;
     try {
-      json = await this.#fileSystem.readFile(keyPath.path, {
+      json = await fs.readFile(keyPath.path, {
         encoding: "utf8",
       });
     } catch (e) {
@@ -301,7 +167,7 @@ export class Web3KeyStore
     const keyPath = await this.#getKeyPath(keyId);
     if (keyPath == null) return { result: "keyNotFound", keyId };
     try {
-      await this.#fileSystem.delete(keyPath.path);
+      await fs.removeFile(keyPath.path);
     } catch (e) {
       return { result: "error", message: `${e}` };
     }
@@ -337,12 +203,12 @@ export class Web3KeyStore
     const keyId = generateKeyId();
     const keyObject = await encryptKeyObject(keyId, privateKey, passphrase);
     try {
-      await this.#fileSystem.mkdir(this.path, { recursive: true });
+      await fs.mkdir(this.path, { recursive: true });
     } catch (e) {
       return { result: "error", message: `${e}` };
     }
     const createdAt = new Date();
-    const keyPath = pathJoin(
+    const keyPath = join(
       this.path,
       `UTC--${createdAt
         .toISOString()
@@ -350,7 +216,7 @@ export class Web3KeyStore
         .replace(/:/g, "-")}--${keyId}`,
     );
     try {
-      await this.#fileSystem.writeFile(
+      await fs.writeFile(
         keyPath,
         JSON.stringify(keyObject),
         "utf8",
