+++ @planetarium/account-web3-secret-storage/src/index.ts
@@ -10,6 +10,4 @@ export {
 export {
   getDefaultWeb3KeyStorePath,
   Web3KeyStore,
-  LocalStorageFileSystem,
-  NodeJsFileSystem,
 } from "./Web3KeyStore.js";
