+++ @planetarium/account-web3-secret-storage/src/path/browser.ts
@@ -1,0 +1,7 @@
+export function join(...args: string[]): string {
+  return args.join("/");
+}
+
+export function getDefaultWeb3KeyStorePath(): string {
+  return "/planetarium/account/web3-secret-storage/keystore";
+}
